fix(arm32): 修复 uart_init 空指针解引用导致控制台波特率被覆盖
现象:内核启动后串口只输出乱码,且 JTAG 心跳里 loop/ticks/irq 全为 0,
看起来像"卡死在 uart_puts"。

根因:重构时把分频搜索挪进 uart_baud_search() 之后,uart_init() 变成
无条件解引用 result:

    uart_baud_search(uart_clk, baud, result);   // result==NULL 时直接 return
    if (result != NULL && result->baudgen == 0) { return; }
    mmio_write32(base + UART_BAUDGEN, result->baudgen);   // 读地址 0

而 uart_baud_search() 对 NULL 是直接返回的,于是 BAUDGEN/BAUDDIV 被写成
从地址 0 读到的值。ARM 关闭 MMU 时地址 0 是 OCM 且可写,所以既不崩溃也
不报错,只是静默配上错误的波特率 —— 极难定位。

system_console.c 的 console_init() 正是用 NULL 调用 uart_init(),
它在 kmain 里紧跟一次正确的 uart_init(&baud_result) 之后执行,
把刚配好的寄存器又覆盖掉了,所以横幅必然乱码。

修复:uart_init() 内部改用局部变量,result 为 NULL 时只做副作用。

顺带修掉两个让定位变难的问题:

1. uart_clk 无法区分"闭环收敛成功"和"收敛失败后用常量兜底" ——
   两者可能返回同一个数字,含义却完全相反。新增 HB[12..15] 记录
   clock_source / 迭代次数 / BAUDGEN / BAUDDIV,并加
   uart_converge_last_iters() 暴露迭代次数。

2. JTAG 验证脚本在 con 之后只等 500ms 就读心跳,而 9600 波特下启动横幅
   本身要 ~870ms,DIRM0/TICKS/IRQCOUNT 这些"横幅之后才写"的槽位必然
   读到 0。已改成轮询 ticks != 0 作为"中断子系统已起来"的门控。
   (OCM 不会被 rst -processor 清零,所以同时加了显式清零防残影。)

实测验证(Zynq-7020, UART1 @ 9600):
  UART ref clk : 100000000 Hz (source=1 iters=1, self-calibrated)
  Baud         : requested=9600 actual=9600 (BAUDGEN=1736 BAUDDIV=5 err=0 ppm)
  IRQ status   : ticks=20 irq_count=20 last_intid=29 spurious=0
  心跳          : ticks=0x2DDF irqcount=0x2DDF   <- 零丢失
  主循环        : Δticks=1611 / Δuptime=1611ms  <- 1kHz 精确

同时把 uart_baud_search() 抽成纯计算模块并补宿主单测:
  tests/test_arm32_uart_baud.py  (BAUDDIV 有效区间 4..254 的回归保护,
                                  该区间曾导致 7.033 倍波特率偏差)
  tests/test_arm32_console.py    (printf 格式化)
为此 types.h 增加 __STDC_HOSTED__ 分支:宿主构建走标准头,
freestanding 构建自带 size_t/uintptr_t/bool 定义。

+++ arch/arm32/src/console.c
@@ -223,7 +223,13 @@ void console_printf(const char *fmt, ...)
             break;
         }
         case 'p': {
-            u32 value = (u32)va_arg(args, void *);
+            /*
+             * 经由 uintptr_t 转换,而不是直接 (u32)va_arg(args, void *)。
+             * 直接截断指针在 32 位 ARM 上恰好能用,但在指针更宽的环境里
+             * 会被编译器判为危险转换(-Wvoid-pointer-to-int-cast)。
+             * 这个写法在两种环境下都表达同样的意图:取指针的数值。
+             */
+            u32 value = (u32)(uintptr_t)va_arg(args, void *);
             console_puts("0x");
             out_unsigned(value, 16, false, 8, true, false);
             fmt++;
